+++ Realm/RLMSchema.mm
@@ -77,6 +77,20 @@ static NSMutableDictionary *s_classNameToMangledName;
     }
 }
 
+static inline bool IsRLMObjectSubclass(Class cls) {
+    Class parent = class_getSuperclass(cls);
+    if (parent != RLMObject.class) {
+        if (parent && RLMIsSubclass(parent, RLMObject.class)) {
+            @throw [NSException exceptionWithName:@"RLMException"
+                                           reason:[NSString stringWithFormat:@"Class '%s' inherits from a RLMObject subclass. This is currently not supported. All model classes must inherit directly from RLMObject and cannot have subclasses of their own.",
+                                                   class_getName(cls)]
+                                         userInfo:nil];
+        }
+        return false;
+    }
+    return true;
+}
+
 + (void)initialize {
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
@@ -89,10 +103,10 @@ static NSMutableDictionary *s_classNameToMangledName;
         unsigned int numClasses;
         Class *classes = objc_copyClassList(&numClasses);
         for (unsigned int i = 0; i < numClasses; i++) {
-            if (class_getSuperclass(classes[i]) != RLMObject.class) {
+            Class cls = classes[i];
+            if (!IsRLMObjectSubclass(cls)) {
                 continue;
             }
-            Class cls = classes[i];
 
             RLMObjectSchema *objectSchema = nil;
 #ifdef REALM_SWIFT